more small tweaks and updates. More rough draft elements
I would say that in its current state this document is quite a bit less polished than the previous commits. I decided to just write down a bunch of stuff though to see what came out, the plan still is to perhaps do a few more drafts and perhaps even migrate this document to LaTeX to be re-posted with new fancy formatting. I am still not sure how serious of a tone I want to go for here, seeing as the whole project is pretty limited in scope and is intended to serve more as a proof of concept than a usable program.